// src/swagger-typescript/model/assetAdministrationShellDescriptor.ts
/**
 * BaSyx Asset Administration Shell HTTP REST-API
 * The full description of the generic BaSyx Asset Administration Shell HTTP REST-API
 *
 * OpenAPI spec version: v1
 * Contact: constantin.ziesche@bosch.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdministrativeInformation } from './administrativeInformation';
import { Asset } from './asset';
import { IEndpoint } from './iEndpoint';
import { Identifier } from './identifier';
import { LangString } from './langString';
import { ModelType } from './modelType';
import { SubmodelDescriptor } from './submodelDescriptor';

export interface AssetAdministrationShellDescriptor { 
    idShort?: string;
    identification?: Identifier;
    administration?: AdministrativeInformation;
    description?: Array<LangString>;
    endpoints?: Array<IEndpoint>;
    readonly category?: string;
    modelType?: ModelType;
    asset?: Asset;
    submodels?: Array<SubmodelDescriptor>;
}